test(core): split session event specs
